src/forecast/requirements.txt
flask
redis
psycopg2-binary
pandas
scikit-learn

